@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,20 +28,20 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  http: ^0.12.2
-  flutter_native_splash: ^1.1.8+4
-  font_awesome_flutter: ^9.0.0
-  firebase_core: ^0.5.3
-  firebase: ^8.0.0
-  cloud_firestore: ^0.14.4
-  firebase_auth: ^0.18.4+1
-  provider: ^4.3.2+1
-  intl: ^0.16.1
+  http: ^0.13.6
+  flutter_native_splash: ^2.3.4
+  font_awesome_flutter: ^10.6.0
+  firebase_core: ^2.19.0
+  firebase: ^9.0.3
+  cloud_firestore: ^4.11.0
+  firebase_auth: ^4.11.1
+  provider: ^6.0.5
+  intl: ^0.18.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.13.1
 
 
 flutter_icons:
