@@ -2,12 +2,12 @@ import 'dart:ui';
 
 import 'package:firebase_auth/firebase_auth.dart';
 
-String userName;
+String? userName;
 int statusIndex = 0;
-String selectedPincode;
+String? selectedPincode;
 Color lightGrey = Color(0xFF344955);
 Color yellow1 = Color(0xFFF9AA33);
 Color darkGrey=Color(0xFF232F34);
 Color paleGrey=Color(0xFF707E86);
-String chosenDate;
-User userDetails;
+String? chosenDate;
+User? userDetails;
